+++ src/oss4climate_app/src/database/__init__.py
@@ -2,14 +2,13 @@
 Module to manage a database input
 """
 
-import os
 from datetime import datetime
 from typing import Optional
 
 import pandas as pd
-from sqlmodel import Field, Session, SQLModel, create_engine
+from sqlmodel import Field, SQLModel
 
-from oss4climate.src.config import SETTINGS
+from oss4climate.src.database import get_engine
 
 
 # -------------------------------------------------------------------------------------
@@ -34,33 +33,10 @@ class SearchLog(SQLModel, table=True):
     view_offset: Optional[int]
 
 
-# -------------------------------------------------------------------------------------
-# Engine and app database connection
-# -------------------------------------------------------------------------------------
-
-
-def _open_engine_and_create_database_if_missing():
-    x = create_engine(
-        SETTINGS.database_connection_string,
-        echo=False,
-    )
-    # TODO : this currently also creates empty tables for the "oss4climate" part of the code,
-    #   this is likely avoidable and could be removed in a later version
-    SQLModel.metadata.create_all(x)
-    return x
-
-
-_ENGINE = _open_engine_and_create_database_if_missing()
-
-
-def open_database_session() -> Session:
-    return Session(_ENGINE)
-
-
 # A quick and dirty dumping of the database as JSON
 def dump_database_request_log_as_csv() -> str:
-    return pd.read_sql_table(RequestLog.__tablename__, _ENGINE).to_csv(index=False)
+    return pd.read_sql_table(RequestLog.__tablename__, get_engine()).to_csv(index=False)
 
 
 def dump_database_search_log_as_csv() -> str:
-    return pd.read_sql_table(SearchLog.__tablename__, _ENGINE).to_csv(index=False)
+    return pd.read_sql_table(SearchLog.__tablename__, get_engine()).to_csv(index=False)
