+++ src/test/oss4climate/test_imports.py
@@ -1,4 +1,5 @@
 from oss4climate.src import models
+from oss4climate.src.database import scrape_cache
 
 
 def test_import():
@@ -6,13 +7,13 @@ def test_import():
 
     import oss4climate
     from oss4climate import src
-    from oss4climate.src import config, database, helpers, log, nlp, parsers
+    from oss4climate.src import config, helpers, log, nlp, parsers
 
     # Package
     assert isinstance(oss4climate.__name__, str)
     assert isinstance(src.__name__, str)
 
-    assert isinstance(database.__name__, str)
+    assert isinstance(scrape_cache.__name__, str)
     assert isinstance(nlp.__name__, str)
     assert isinstance(parsers.__name__, str)
     assert isinstance(config.__name__, str)
